+++ .gitignore
@@ -4,6 +4,8 @@ build/
 !gradle/wrapper/gradle-wrapper.jar
 !**/src/main/**/build/
 !**/src/test/**/build/
+**/application.yml
+.DS_Store
 
 ### STS ###
 .apt_generated
