+++ src/main/java/org/yourssu/festa/controller/BoothController.java
@@ -1,0 +1,11 @@
+package org.yourssu.festa.controller;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/api/booths")
+public class BoothController {
+}
