+++ src/main/java/org/yourssu/festa/domain/Product.java
@@ -1,5 +1,7 @@
 package org.yourssu.festa.domain;
 
+import org.yourssu.festa.domain.entity.ProductEntity;
+
 public record Product(
         Long id,
         String name,
