+++ src/main/java/org/yourssu/festa/domain/entity/ProductEntity.java
@@ -1,15 +1,30 @@
-package org.yourssu.festa.domain;
+package org.yourssu.festa.domain.entity;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
+import lombok.Getter;
 import lombok.NoArgsConstructor;
+import org.hibernate.annotations.ColumnDefault;
 
 @Entity
-@Table(name = "booth_img")
+@Table(name = "product")
+@Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class BoothImageEntity {
+public class ProductEntity {
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
+
+    @Column
+    private String name;
+
+    @Column
+    private int price;
+
+    @Column
+    private String description;
+
+    @ColumnDefault("false")
+    private boolean soldout;
 
     @Column(name = "img_url")
     private String imgUrl;
