+++ src/main/java/org/yourssu/festa/reader/BoothReader.java
@@ -1,11 +1,10 @@
-package org.yourssu.festa.service;
+package org.yourssu.festa.reader;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import org.yourssu.festa.common.code.AuthErrorCode;
 import org.yourssu.festa.common.code.BoothErrorCode;
-import org.yourssu.festa.domain.Booth;
-import org.yourssu.festa.domain.BoothEntity;
+import org.yourssu.festa.domain.entity.BoothEntity;
 import org.yourssu.festa.exception.CustomException;
 import org.yourssu.festa.repository.BoothRepository;
